Convert some straggler multiversion tests to use the new multiversion base class
