--- subprojects/testing-jvm/src/integTest/resources/org/gradle/testing/IncrementalTestIntegrationTest/doesNotRunStaleTests/src/test/java/Broken.java
@@ -1,8 +1,0 @@
-import org.junit.Test;
-
-public class Broken {
-    @Test
-    public void broken() {
-        throw new RuntimeException("broken");
-    }
-}
