--- subprojects/testing-jvm/src/integTest/resources/org/gradle/testing/IncrementalTestIntegrationTest/executesTestsWhenSourceChanges/NewMainClass.java
@@ -1,6 +1,0 @@
-public class MainClass {
-    @Override
-    public String toString() {
-        return "new main class";
-    }
-}
