--- subprojects/testing-jvm/src/integTest/resources/org/gradle/testing/IncrementalTestIntegrationTest/executesTestsWhenSourceChanges/NewOk.java
@@ -1,5 +1,0 @@
-public class Ok {
-    @org.junit.Test
-    public void someTest() {
-    }
-}
