--- subprojects/testing-jvm/src/integTest/resources/org/gradle/testing/IncrementalTestIntegrationTest/executesTestsWhenSourceChanges/src/main/java/MainClass.java
@@ -1,2 +1,0 @@
-class MainClass {
-}
