--- subprojects/testing-jvm/src/integTest/resources/org/gradle/testing/IncrementalTestIntegrationTest/executesTestsWhenTestFrameworkChanges/build.gradle
@@ -1,14 +1,0 @@
-apply plugin: 'java'
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    testImplementation 'junit:junit:4.13'
-    testImplementation 'org.testng:testng:6.3.1'
-}
-
-test {
-    include '**/*Test.*'
-}
