--- subprojects/testing-jvm/src/integTest/resources/org/gradle/testing/IncrementalTestIntegrationTest/executesTestsWhenTestFrameworkChanges/src/test/java/JUnitExtra.java
@@ -1,7 +1,0 @@
-import org.junit.Test;
-
-public class JUnitExtra {
-    @Test
-    public void ok() {
-    }
-}
