--- subprojects/testing-jvm/src/integTest/resources/org/gradle/testing/IncrementalTestIntegrationTest/executesTestsWhenTestFrameworkChanges/src/test/java/JUnitTest.java
@@ -1,7 +1,0 @@
-import org.junit.Test;
-
-public class JUnitTest {
-    @Test
-    public void ok() {
-    }
-}
