--- subprojects/testing-jvm/src/integTest/resources/org/gradle/testing/IncrementalTestIntegrationTest/executesTestsWhenTestFrameworkChanges/src/test/java/TestNGTest.java
@@ -1,6 +1,0 @@
-import org.testng.annotations.Test;
-
-public class TestNGTest {
-    @Test
-    public void ok() {}
-}
