--- subprojects/testing-jvm/src/integTest/resources/org/gradle/testing/IncrementalTestIntegrationTest/shared/src/test/java/Ok.java
@@ -1,7 +1,0 @@
-import org.junit.Test;
-
-public class Ok {
-    @Test
-    public void ok() {
-    }
-}
